Update ssr-and-ssg.md
I noticed you start off explaining what SSR is without clarifying what the acronym stands for, so I've added a few hints here and there. Adding some extra keywords for SEO can't hurt either. I've also added a clarification around SSG to make it clear that it produces actual static html pages that can be hosted on a CDN, as the original reference to "running" on any server suggests a runtime, which can be misleading. Perhaps consider also writing out the acronyms in the menu.